manager/requirements.txt
Django==5.0.2
daphne==4.1.0
django-cors-headers==4.3.1
rich==13.7.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-rest-swagger==2.2.0
drf-yasg==1.21.7
redis==5.0.2
django-redis==5.4.0
channels==4.0.0
schedule==1.2.1
tzdata==2024.1
channels-redis==4.2.0
drf-spectacular==0.27.1
django-constance==3.1.0
django-rich-logging==0.3.0
django-filter==24.1
pydantic==2.6.4